Last used controls get reset when forms are added/removed
Extra precaution although highly unlikely that a circumstance would occur where this is useful. One such circumstance is when a DropButtonClick event for a combobox  is triggered while the combo is already expanded (a second time)  and it opens a new modal form which would not scroll due to the fact that the last cached combo is from a different form and is still expanded. Again, unlikely but better to have the safety in place