Removed swapping MouseProc signature for x64 stack bug
The previous assumption that we are swapping delegates was simply wrong. The code was swapping bytes in assembly and it was only swapping the 5 lower bytes of the PCode address while also swapping 3 bytes of instructions. The correct offset for swaping an 8-byte pointer to PCode would be 55 instead of 52. However, this is unnecessary because the code now simply replaces RET 0x18 with RET 0x00 which basically means that it does not remove the 24 bytes used by the arguments.